move TxTest files to test/ directory
