+++ src/main/CommandHandler.cpp
@@ -27,7 +27,7 @@
 #include "ExternalQueue.h"
 
 #include <regex>
-#include "transactions/TxTests.h"
+#include "test/TxTests.h"
 using namespace stellar::txtest;
 
 using std::placeholders::_1;
