+++ src/transactions/OfferTests.cpp
@@ -9,8 +9,8 @@
 #include "main/test.h"
 #include "lib/catch.hpp"
 #include "util/Logging.h"
-#include "TxTests.h"
 #include "util/TestUtils.h"
+#include "test/TxTests.h"
 #include "util/Timer.h"
 #include "database/Database.h"
 #include "ledger/LedgerDelta.h"
