+++ src/transactions/SetOptionsTests.cpp
@@ -12,7 +12,7 @@
 #include "lib/catch.hpp"
 #include "util/Logging.h"
 #include "util/TestUtils.h"
-#include "TxTests.h"
+#include "test/TxTests.h"
 #include "transactions/TransactionFrame.h"
 #include "ledger/LedgerDelta.h"
